Исправления

+++ MyQFileDir.h
@@ -23,7 +23,7 @@ struct MyQFileDir
     enum { modified = 1, read = 2, noSort = 0 };
     inline static QString RemoveOldFiles(QString directory, int remainCount, int sortFlag = MyQFileDir::modified);
 
-    inline static void RemoveOldFiles2(const QString &backupRootDir, int maxDays);
+    inline static void RemoveOldFiles2(const QString &directory, int maxDays);
 
     inline static void RemoveEmptySubcats(const QString &directory);
 
@@ -135,60 +135,44 @@ QString MyQFileDir::RemoveOldFiles(QString directory, int remainCount, int sortF
     return ret;
 }
 
-void MyQFileDir::RemoveOldFiles2(const QString &backupRootDir, int maxDays)
+void MyQFileDir::RemoveOldFiles2(const QString &directory, int maxDays)
 {
-    QDir dir(backupRootDir);
-    if (!dir.exists()) {
+    QDir dir(directory);
+    if (!dir.exists())
+    {
 	return;
     }
 
-    QDateTime now = QDateTime::currentDateTime();
+    QDateTime currentDate = QDateTime::currentDateTime();
 
-    QStringList subDirs = dir.entryList(QDir::Dirs | QDir::NoDotAndDotDot);
-    for (const QString &subDirName : subDirs) {
-	// Парсим имя папки: "yyyy.MM.dd hh-mm-ss_mode"
-	int firstSpace = subDirName.indexOf(' ');
-	if (firstSpace == -1) {
-	    continue;
-	}
+    QDirIterator it(directory, QDir::Files | QDir::NoDotAndDotDot, QDirIterator::Subdirectories);
+    while(it.hasNext()) {
+	it.next();
+	QFileInfo fileInfo = it.fileInfo();
+	QDateTime lastModified = fileInfo.lastModified();
+	qint64 daysDiffrence = lastModified.daysTo(currentDate);
 
-	QString datePart = subDirName.left(firstSpace);
-	QString timeAndModePart = subDirName.mid(firstSpace + 1);
-
-	int firstUnderscore = timeAndModePart.indexOf('_');
-	if (firstUnderscore == -1) {
-	    continue;
-	}
-
-	QString timePart = timeAndModePart.left(firstUnderscore);
-	QString dateTimeStr = datePart + " " + timePart;
-
-	QDateTime dirDateTime = QDateTime::fromString(dateTimeStr, "yyyy.MM.dd hh-mm-ss");
-	if (!dirDateTime.isValid()) {
-	    continue;
-	}
-
-	qint64 daysAgo = dirDateTime.daysTo(now);
-	if (daysAgo > maxDays) {
-	    QDir subDirFullPath(dir.filePath(subDirName));
-	    subDirFullPath.removeRecursively();
+	if(daysDiffrence > maxDays)
+	{
+	    QFile::remove(fileInfo.absoluteFilePath());
 	}
     }
 
-    RemoveEmptySubcats(backupRootDir);
+    RemoveEmptySubcats(directory);
 }
 
 void MyQFileDir::RemoveEmptySubcats(const QString &directory)
 {
-        QDir dir(directory);
+    QDir dir(directory);
+    // Рекурсивное обработка поддиректорий
+    const QStringList subDirs = dir.entryList(QDir::Dirs | QDir::NoDotAndDotDot);
+    for (const QString &subDir : subDirs)
+    {
+	RemoveEmptySubcats(dir.filePath(subDir));
+    }
 
-	// Рекурсивное обработка поддиректорий
-	const QStringList subDirs = dir.entryList(QDir::Dirs | QDir::NoDotAndDotDot);
-	for (const QString &subDir : subDirs) {
-	        RemoveEmptySubcats(dir.filePath(subDir));
-	}
-
-	// Рекурсивное удаление пустых папок
+    // удаление пустой папки
+    if(dir.isEmpty())
 	dir.rmdir(".");
 }
